Todo Funcional

+++ src/main/java/org/iesvdm/dao/ClienteDAO.java
@@ -14,6 +14,6 @@ public interface ClienteDAO {
 	
 	public void update(Cliente cliente);
 	
-	public void delete(long id);
+	public void delete(int id);
 	
 }
